requirements.txt
Django==3.2.0
django-environ==0.4.5
djangorestframework==3.11.0
python-dateutil==2.8.1
virtualenv==20.0.10
virtualenvwrapper-win==1.2.6
gunicorn == 20.0.4
django-cors-headers==3.7.0
# Dev-Requirements
pylint-django
autopep8
ruamel.yaml